requirements.txt
SQLAlchemy>=2.0
python-dotenv
mysql-connector-python
PyMySQL
